Split include/sources
