+++ include/mc_state_observation/AttitudeObserver.h
